app/backend/requirements-backend.txt
# Backend Application Dependencies
# Lightweight libraries for API, inference, and production use

# ========================
# Web Framework
# ========================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ========================
# Database
# ========================
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9  # PostgreSQL adapter
asyncpg==0.29.0  # Async PostgreSQL

# ========================
# Authentication & Security
# ========================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pydantic[email]==2.5.2
pydantic-settings==2.1.0

# ========================
# AI & Agents
# ========================
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langgraph==0.0.20
langsmith==0.0.77
groq==0.4.1  # Groq API for LLMs
google-generativeai==0.3.2  # Gemini API

# ========================
# ML Inference (Lightweight)
# ========================
onnxruntime==1.16.3  # ONNX inference (CPU)
# onnxruntime-gpu==1.16.3  # Use this for GPU inference
tensorflow-lite==2.15.0  # TensorFlow Lite for inference
scikit-learn==1.3.2  # For XGBoost models
xgboost==2.0.2
numpy==1.24.3
opencv-python-headless==4.8.1.78  # Headless OpenCV (no GUI)
Pillow==10.1.0

# ========================
# Task Queue & Caching
# ========================
celery==5.3.4
redis==5.0.1
flower==2.0.1  # Celery monitoring

# ========================
# HTTP Client
# ========================
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# ========================
# Data Processing
# ========================
pandas==2.1.3
geopandas==0.14.1  # Lightweight geospatial
shapely==2.0.2
pyproj==3.6.1

# ========================
# Web Scraping (for news)
# ========================
beautifulsoup4==4.12.2
newspaper3k==0.2.8
feedparser==6.0.10

# ========================
# Utilities
# ========================
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
python-dateutil==2.8.2
pytz==2023.3

# ========================
# Monitoring & Observability
# ========================
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.1

# ========================
# CORS & Middleware
# ========================
python-cors==1.0.0

# ========================
# Testing
# ========================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI
faker==20.1.0  # Generate fake data for tests
